+++ lib/pages/setup/workspace_selection_page.dart
@@ -17,6 +17,7 @@ import 'package:url_launcher/url_launcher_string.dart';
 
 import '../../ui/back_button.dart';
 import '../../ui/custom_dropdown.dart';
+import '../../ui/custom_scaffold.dart';
 import '../../ui/dropdown_button3.dart';
 import '../../ui/widgets.dart';
 import 'project_selection_page.dart';
@@ -57,7 +58,7 @@ class _WorkspaceSelectionPageState extends State<WorkspaceSelectionPage> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return CustomScaffold(
       body: Center(
         child: Padding(
           padding: const EdgeInsets.all(32),